Fixes #2 intersection/2 would compute incorrect interval for unbounded
Since `min_endpoint/2` and `max_endpoint/2` was used (which always prefers
to return `:unbounded`, if any endpoint was unbounded, then that
would get picked.
This is the behaviour we want for `union/2`, but not for `intersection/2`.
I've added a third argument to `min_endpoint/2` and `max_endpoint/2` so that
they now accept `:prefer_unbounded` or `:prefer_bounded`.

I'd like to rewrite intersection and union, so I don't want to spend
too much time on the beauty of this code right now, as long as it is correct.