Adding `Interval.size/2`
Also adding:

- `Interval.Point.subtract/3`
- `Interval.Point.add/2`

Currently, `size/2` is reporting empty intervals as `nil` size,
because we don't know what point type the interval is,
because we represent an empty interval as `left: :empty, right: :empty`.

If we _didn't_ do this, empty intervals would not have this problem.
If we instead normalized all empty intervals to

```elixir
[left: point, right: point, bounds: "()"]
```

it would be almost as cheap to check for empty as it is now,
but we'd still have a point to reference when doing size checks.

I think this might be a better solution, but we need to overhaul
the interval and fix all of the `:empty` handling.