Remove the hack that an empty interval uses `:empty` atom
Instead, a normalized empty interval is `(zero,zero)`, where `zero` is
a new concept that each `Point` implementation must provide.

A Point protocol function `zero(a)` returns the "zero" value for
the given Point type, which we use as the point in the empty interval
(both left and right).

This allows us to define `size/2` so that an empty and non-empty interval
is calculated in the same way, which is way more elegant.

It also allows us to maintain a reference to the point implementation
even when the interval becomes empty, which can be helpful in the future.

Functionally, there is no changes to the API.
The empty interval is now just represented differently.