Interval coverage, now 100%
Had to remove some code that is never really used because when
properly normalized, it isn't relevant.

Replaced by code that asserts proper normalization,
and a single test case that touches that code to ensure it works.

Also found a bug in the resulting bounds when an intersection
is calculated.
I'd like to reproduce that with property tests so I know that those
will cover that kind of error.