Adding Interval inference implementation
It's now possible (for the defined value types) to infer what interval type
to use, as long as either the left or right side is a non-nil value:

```
Interval.new(left: 1, right: 2)
```

Is now equivalent to

```
Interval.Integer.new(left: 1, right: 2)
```

This won't work if the interval is both left and right unbounded.

```
Interval.Integer.new(left: 1, right: 2)
```

is still the preferred way to construct an interval when you know
what interval implementation you want to use.