fix(contains): fix issue #29 contains?/2 bug
This commit fixes a bug in the `Interval.contains?/2` that caused an
interval with identical points but different bounds to not be considered
contained when it should have been.

E.g. `[1,2]` did not contain `(1,2)`.

This commit adds a regression test for the bug fixed in this commit.

This commit also adds `Interval.exclusive_left?/1` and `Interval.exclusive_right?/1`
(previously we only had `Interval.inclusive_left?/1` and `Interval.inclusive_right?/1`).

This commit alters the behavior of `Interval.contains?/2` in a backwards
incompatible way. `contains?/2` now returns true for any interval if
the `b` interval is empty.

This was previously not the case, however this behaviour I now consider
a bug, because it breaks the "does A contains all points in B" definition.
This also aligns more closely with what Postgres does, which is the target.

This commit also raises Credo's cyclomatic complexity limit to 22,
because rewriting `Interval.contains?/2` to be more readable is not
a priority right now and I want this in and released :)